fix: refactor timeline
